Switch vclock implementation to maps
Execution overhead is a bit higher, and we've seen in benchmarks that
the `orddict` implementation is the fastest since the order is stable.
However, the memory usage for an orddict is quite high, at least in
comparison with maps. In the commit queue, we are able to reduce memory
usage by 30% just by switching the clock implemenation to use maps.
Further usage of maps inside the commit queue yields negligible
benefits.

TODO: measure execution overhead against baseline